# server/apidocs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Beskar API",
    "version": "1.0.0",
    "description": "Authentication and implemented whiteboarding v2 APIs.\n\n**Try it:** [Sign in](/auth/login?returnTo=%2Fapi%2Fdocs), then use the browser-session option, or select BearerToken and paste a valid Zitadel JWT (without the Bearer prefix). The server currently verifies bearer JWTs using its OIDC ID-token verifier; opaque access tokens are not supported. Calls run against this origin with your permissions and can create real data.\n\nReplace example space/parent IDs with your own. Generate an Idempotency-Key UUID for each new board or checkpoint batch and keep it with the same payload for retries. Login, registration and logout require browser navigation. Cookie sessions normally require HTTPS.",
    "license": {
      "name": "Apache 2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "This Beskar instance"
    }
  ],
...
  "security": [
    {
      "BrowserSession": []
    },
    {
      "BearerToken": []
    }
  ],
  "paths": {
...
    "/api/v2/editor/space/{spaceId}/whiteboard/create": {
      "post": {
        "tags": [
          "Whiteboarding v2"
        ],
        "operationId": "createWhiteboardV2",
        "summary": "Create a whiteboard",
        "description": "Creates a board in the v2 schema. Requires space edit_page permission and parent edit permission when nested. Generate a fresh UUID Idempotency-Key per create action; reuse it and the same body on retries. Replays return 201 with the original IDs. Unknown fields, invalid UTF-8 and trailing JSON are rejected. Maximum body: 16 KiB. V2 incremental checkpoints are supported. V2 retrieval and publication are not implemented yet; v1 readers cannot load this board.",
        "parameters": [
          {
            "name": "spaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Nonzero UUID."
            },
            "example": "c7022348-1bb3-4e52-8403-17786e15e035"
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Nonzero UUID."
            },
            "description": "Exactly one nonzero UUID. Generate with crypto.randomUUID(); retain for retries, replace for each new create action."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWhiteboardRequest"
              },
              "examples": {
                "topLevel": {
                  "summary": "Top-level board",
                  "value": {
                    "title": "Architecture discussion"
                  }
                },
                "nested": {
                  "summary": "Under an existing parent",
                  "value": {
                    "title": "Architecture discussion",
                    "parentId": 123
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created, or replayed successfully with the same page and space IDs.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateWhiteboardResponse"
                }
              }
            }
          },
          "302": {
            "description": "Browser redirect. Follow the Location header through the identity provider.",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "Redirect destination."
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED from the handler, or an authentication middleware error with a numeric code.",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/WhiteboardError"
                    },
                    {
                      "$ref": "#/components/schemas/AuthenticationError"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid space UUID, Idempotency-Key, or JSON body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                },
                "example": {
                  "status": "FAILED",
                  "error": {
                    "code": "INVALID_REQUEST",
                    "message": "Invalid space UUID, Idempotency-Key, or JSON body."
                  }
                }
              }
            }
          },
          "403": {
            "description": "Space edit_page or parent edit permission denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                },
                "example": {
                  "status": "FAILED",
                  "error": {
                    "code": "CREATE_WHITEBOARD_FORBIDDEN",
                    "message": "Space edit_page or parent edit permission denied."
                  }
                }
              }
            }
          },
          "404": {
            "description": "Space or parent missing/deleted, or parent belongs to another space.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                },
                "example": {
                  "status": "FAILED",
                  "error": {
                    "code": "SPACE_OR_PARENT_NOT_FOUND",
                    "message": "Space or parent missing/deleted, or parent belongs to another space."
                  }
                }
              }
            }
          },
          "409": {
            "description": "IDEMPOTENCY_KEY_REUSED: changed content with an existing key. SPACE_ARCHIVED: space is archived.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                },
                "example": {
                  "status": "FAILED",
                  "error": {
                    "code": "IDEMPOTENCY_KEY_REUSED",
                    "message": "IDEMPOTENCY_KEY_REUSED: changed content with an existing key. SPACE_ARCHIVED: space is archived."
                  }
                }
              }
            }
          },
          "413": {
            "description": "JSON body exceeds 16 KiB.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                },
                "example": {
                  "status": "FAILED",
                  "error": {
                    "code": "REQUEST_TOO_LARGE",
                    "message": "JSON body exceeds 16 KiB."
                  }
                }
              }
            }
          },
          "415": {
            "description": "Content-Type must be application/json.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                },
                "example": {
                  "status": "FAILED",
                  "error": {
                    "code": "UNSUPPORTED_MEDIA_TYPE",
                    "message": "Content-Type must be application/json."
                  }
                }
              }
            }
          },
          "500": {
            "description": "Creation could not be confirmed. Retry the same request and key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                },
                "example": {
                  "status": "FAILED",
                  "error": {
                    "code": "WHITEBOARD_CREATE_FAILED",
                    "message": "Creation could not be confirmed. Retry the same request and key."
                  }
                }
              }
            }
          },
          "503": {
            "description": "WHITEBOARD_CREATE_BUSY: lock timeout. WHITEBOARD_PERMISSIONS_PENDING: creation recorded but permission setup pending. Retry the same request and key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhiteboardError"
                },
                "example": {
                  "status": "FAILED",
                  "error": {
                    "code": "WHITEBOARD_CREATE_BUSY",
                    "message": "WHITEBOARD_CREATE_BUSY: lock timeout. WHITEBOARD_PERMISSIONS_PENDING: creation recorded but permission setup pending. Retry the same request and key."
                  }
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "1"
                  ]
                },
                "description": "Retry after one second."
              }
            }
          }
        }
      }
    },
...
    "/api/v1/editor/space/{spaceId}/page/{pageId}/metadata": {
      "get": {
        "tags": [
          "Page navigation"
        ],
        "operationId": "getSharedPageMetadata",
        "summary": "Resolve page content API navigation",
        "description": "Requires page view permission. Unpublished v2 boards are visible only to editors. Editors see current draft titles; other viewers see published titles. Deleted spaces are excluded. Archived spaces remain readable under existing page permissions. These shared endpoints remain v1; use contentApiVersion and whiteboard URLs to select v2 content APIs.",
        "parameters": [
          {
            "name": "spaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "pageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page discovery result.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "string",
                      "const": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/SharedPageMetadata"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Authentication or permission denied."
          },
          "404": {
            "description": "Page not found or unavailable to this viewer."
          },
          "500": {
            "description": "Permission service or database error."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Valid Zitadel JWT accepted by the server OIDC verifier. Paste only the token."
      },
      "BrowserSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "zitadel.session",
        "description": "Sign in using the link above. The browser sends the HttpOnly session cookie automatically; do not paste a cookie value."
      }
    },
    "schemas": {
      "CreateWhiteboardRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "title"
        ],
        "properties": {
          "title": {
            "type": "string",
            "description": "Trimmed before validation; 1\u2013255 Unicode characters after trimming. NUL characters are rejected.",
            "example": "Architecture discussion"
          },
          "parentId": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 1,
            "maximum": 9007199254740991,
            "description": "Parent page in this space. Omit or use null for a top-level board."
          }
        }
      },
      "CreateWhiteboardResponse": {
        "type": "object",
        "required": [
          "status",
          "data"
        ],
        "properties": {
          "status": {
            "const": "success"
          },
          "data": {
            "type": "object",
            "required": [
              "pageId",
              "spaceId"
            ],
            "properties": {
              "pageId": {
                "type": "integer",
                "format": "int64",
                "example": 42
              },
              "spaceId": {
                "type": "string",
                "format": "uuid",
                "description": "Nonzero UUID.",
                "example": "c7022348-1bb3-4e52-8403-17786e15e035"
              }
            }
          }
        }
      },
      "WhiteboardError": {
        "type": "object",
        "required": [
          "status",
          "error"
        ],
        "properties": {
          "status": {
            "const": "FAILED"
          },
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  }
}